Support both JSON and Protobuf for VehiclePositions
