+++ apps/parse/lib/parse/vehicle_positions.ex
@@ -1,54 +1,42 @@
 defmodule Parse.VehiclePositions do
   @moduledoc """
 
-  Parses an enhanced Vehicle Position JSON file into a list of `%Model.Vehicle{}` structs.
+  Parser for the VehiclePositions.pb GTFS-RT file.
 
   """
   @behaviour Parse
   alias Model.Vehicle
+  alias Parse.Realtime.FeedMessage
   import Parse.Helpers
 
-  @impl true
-  def parse(body) do
-    body
-    |> Jason.decode!()
-    |> Map.get("entity")
-    |> Enum.flat_map(&parse_entity/1)
+  def parse("{" <> _ = blob) do
+    Parse.VehiclePositionsJson.parse(blob)
   end
 
-  def parse_entity(
-        %{
-          "vehicle" => %{
-            "position" => position,
-            "trip" => trip,
-            "vehicle" => vehicle,
-            "timestamp" => timestamp
-          }
-        } = entity
-      ) do
-    data = Map.get(entity, "vehicle")
-
-    [
-      %Vehicle{
-        id: optional_field_copy(vehicle, "id"),
-        trip_id: optional_field_copy(trip, "trip_id"),
-        route_id: optional_field_copy(trip, "route_id"),
-        direction_id: optional_field_copy(trip, "direction_id"),
-        stop_id: Map.get(data, "stop_id"),
-        label: optional_field_copy(vehicle, "label"),
-        latitude: optional_field_copy(position, "latitude"),
-        longitude: optional_field_copy(position, "longitude"),
-        bearing: optional_field_copy(position, "bearing"),
-        speed: optional_field_copy(position, "speed"),
-        current_status: parse_status(Map.get(data, "current_status")),
-        current_stop_sequence: Map.get(data, "current_stop_sequence"),
-        updated_at: unix_to_local(timestamp)
-      }
-    ]
+  def parse(blob) do
+    blob
+    |> FeedMessage.decode()
+    |> (fn message -> message.entity end).()
+    |> Stream.map(fn entity -> entity.vehicle end)
+    |> Stream.map(&parse_vehicle_update/1)
   end
 
-  def parse_entity(%{}) do
-    []
+  def parse_vehicle_update(update) do
+    %Vehicle{
+      id: optional_field_copy(update.vehicle, :id),
+      trip_id: optional_field_copy(update.trip, :trip_id),
+      route_id: optional_field_copy(update.trip, :route_id),
+      direction_id: update.trip && update.trip.direction_id,
+      stop_id: optional_copy(update.stop_id),
+      label: optional_field_copy(update.vehicle, :label),
+      latitude: update.position && update.position.latitude,
+      longitude: update.position && update.position.longitude,
+      bearing: update.position && update.position.bearing,
+      speed: update.position && update.position.speed,
+      current_status: current_status(update.current_status),
+      current_stop_sequence: update.current_stop_sequence,
+      updated_at: unix_to_local(update.timestamp)
+    }
   end
 
   defp optional_copy("") do
@@ -69,19 +57,19 @@ defmodule Parse.VehiclePositions do
     nil
   end
 
-  defp parse_status(nil) do
+  defp current_status(nil) do
     :in_transit_to
   end
 
-  defp parse_status("IN_TRANSIT_TO") do
+  defp current_status(:IN_TRANSIT_TO) do
     :in_transit_to
   end
 
-  defp parse_status("INCOMING_AT") do
+  defp current_status(:INCOMING_AT) do
     :incoming_at
   end
 
-  defp parse_status("STOPPED_AT") do
+  defp current_status(:STOPPED_AT) do
     :stopped_at
   end
